Moved more of the capture related types into closure.rs
